useEffect処理を正しくマウントできるように修正
