+++ frontend/twitter-clone/src/components/pages/Home.tsx
@@ -3,14 +3,26 @@ import TweetBox from "./TweetBox";
 import TweetList from "./TweetList";
 import { useEffect, useState } from "react";
 import { useTweetContext } from "../../contexts/TweetProvider";
+import tweetApi from "../../api/tweetApi";
 
 const Home = () => {
-	const { tweets } = useTweetContext();
+	const { tweets, setTweets } = useTweetContext();
 	const [loading, setLoading] = useState<boolean>(true);
 
 	useEffect(() => {
-		setLoading(false);
-	}, [tweets]);
+		const getTweets = async () => {
+			try {
+				const res = await tweetApi.search();
+				setTweets(res.data);
+				setLoading(false);
+			} catch (err: any) {
+				const errors = err.data.errors;
+				console.log(errors);
+				setLoading(false);
+			}
+		};
+		getTweets();
+	}, [setTweets]);
 
 	return (
 		<Box sx={{ height: "100vh", margin: "10px 10px 0", maxWidth: 500 }}>
