+++ frontend/twitter-clone/src/contexts/TweetProvider.tsx
@@ -24,19 +24,6 @@ type TweetProviderProps = {
 export const TweetProvider = ({ children }: TweetProviderProps) => {
 	const [tweets, setTweets] = useState<Tweet[]>([]);
 
-	useEffect(() => {
-		const getTweets = async () => {
-			try {
-				const res = await tweetApi.search();
-				setTweets(res.data);
-			} catch (err: any) {
-				const errors = err.data.errors;
-				console.log(errors);
-			}
-		};
-		getTweets();
-	}, []);
-
 	return (
 		<TweetContext.Provider value={{ tweets, setTweets }}>
 			{children}
